pyproject.toml
[tool.poetry]
name = "outlinerobot"
version = "0.1.0"
description = "A Telegram bot for outline.com"
authors = ["Antonio Pitasi <me@anto.pt>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^12.6.1"
requests = "^2.23.0"
cachetools = "^4.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
